requirements.txt
ccxt>=4.4.0
python-telegram-bot==20.0
pandas
pandas_ta
python-dotenv
flask
gunicorn
pyyaml

